requirements.txt
# RASC Requirements
# Python 3.8+

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
ultralytics>=8.0.0  # YOLOv8
opencv-python>=4.8.0
Pillow>=10.0.0

# NLP
transformers>=4.30.0
datasets>=2.12.0
sentencepiece>=0.1.99

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Metrics
scikit-learn>=1.3.0
pycocoevalcap>=1.2  # For CIDEr and other captioning metrics

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Logging & Experiment Tracking
tensorboard>=2.13.0
wandb>=0.15.0  # Optional

# Utilities
python-dateutil>=2.8.2
requests>=2.31.0

# Development
jupyter>=1.0.0
ipykernel>=6.23.0
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Documentation
sphinx>=6.2.0
sphinx-rtd-theme>=1.2.0
